pyproject.toml
[project]
name = "classifier-assignment"
version = "0.1.0"
description = "Classifier"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
   "fastapi>=0.115.12",
   "google-genai>=1.16.1",
   "python-dotenv>=1.1.0",
   "python-multipart>=0.0.20",
   "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.5",
]
